move actorForRole into separate intent module
